requirements.txt
aiogram==2.25.1
aiosignal==1.3.1
amocrm-api==2.5.0
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
